Add example for our services
